Halifax lakes water chemistry and top/bottom diatom counts

A subset of well-labeled, clean diatom count data for 44 Halifax-area (Nova Scotia) lakes, an analysis of which has been published by Ginn et al. (2015). These data are in the form of a mudata object, from which location information, parameter information, and data can be extracted using tbl_locations, tbl_params, and tbl_data (respectively). Also, a subset of these data amenable to plotting have been made available.

halifax_lakes

halifax_lakes_plottable

Format

An object of class mudata (inherits from list) of length 5.

Source

Neotoma paleoecology database (http://neotomadb.org/)

References

Ginn, Brian K., Thiyake Rajaratnam, Brian F. Cumming, and John P. Smol. "Establishing Realistic Management Objectives for Urban Lakes Using Paleolimnological Techniques: An Example from Halifax Region (Nova Scotia, Canada)." Lake and Reservoir Management 31, no. 2 (April 3, 2015): 92-108. https://doi.org/10.1080/10402381.2015.1013648.

Examples

library(mudata2)
data(halifax_lakes)

halifax_lakes %>%
  tbl_data()

#> # A tibble: 4,370 x 11
#>    dataset location param sample_type    value neotoma_sample_… depth unit_name
#>    <chr>   <chr>    <chr> <chr>          <dbl>            <int> <dbl> <chr>    
#>  1 neotom… Anderso… Alka… water chem… -0.24              166624    NA Anderson…
#>  2 neotom… Anderso… Alum… water chem…  0.12              166624    NA Anderson…
#>  3 neotom… Anderso… Ammo… water chem…  0.01              166624    NA Anderson…
#>  4 neotom… Anderso… Calc… water chem…  2                 166624    NA Anderson…
#>  5 neotom… Anderso… Chlo… water chem… 12.1               166624    NA Anderson…
#>  6 neotom… Anderso… Chlo… water chem…  0.00456           166624    NA Anderson…
#>  7 neotom… Anderso… Diss… water chem…  3.6               166624    NA Anderson…
#>  8 neotom… Anderso… Iron  water chem…  0.07              166624    NA Anderson…
#>  9 neotom… Anderso… Magn… water chem…  0.61              166624    NA Anderson…
#> 10 neotom… Anderso… Mang… water chem…  0.05              166624    NA Anderson…
#> # … with 4,360 more rows, and 3 more variables: neotoma_dataset_id <dbl>,
#> #   variable_units <chr>, dataset_type <chr>