Nested Principal Components Analysis (PCA)

Powered by prcomp. When creating the nested_data, the data should be scaled (i.e, trans = scale) if all variables are not in the same unit.

nested_prcomp(.data, data_column = .data$data, ...)

Arguments

.data
A data frame with a list column of data frames, possibly created using nested_data.

data_column
An expression that evalulates to the data object within each row of .data

...
Passed to prcomp.

Value

.data with additional columns 'model', 'loadings', 'variance' and 'scores'

Examples

library(tidyr)

nested_pca <- alta_lake_geochem %>%
  nested_data(
    qualifiers = c(depth, zone),
    key = param,
    value = value,
    trans = scale
  ) %>%
  nested_prcomp()

# get variance info
nested_pca %>% unnest(variance)

#> # A tibble: 6 x 13
#>   discarded_colum… discarded_rows qualifiers data  model component
#>   <list>           <list>         <list>     <lis> <lis>     <int>
#> 1 <tibble [32 × 0… <tibble [0 × … <tibble [… <tib… <prc…         1
#> 2 <tibble [32 × 0… <tibble [0 × … <tibble [… <tib… <prc…         2
#> 3 <tibble [32 × 0… <tibble [0 × … <tibble [… <tib… <prc…         3
#> 4 <tibble [32 × 0… <tibble [0 × … <tibble [… <tib… <prc…         4
#> 5 <tibble [32 × 0… <tibble [0 × … <tibble [… <tib… <prc…         5
#> 6 <tibble [32 × 0… <tibble [0 × … <tibble [… <tib… <prc…         6
#> # … with 7 more variables: component_text <chr>, standard_deviation <dbl>,
#> #   variance <dbl>, variance_proportion <dbl>,
#> #   variance_proportion_cumulative <dbl>, loadings <list>, scores <list>

# get loadings info
nested_pca %>% unnest(loadings)

#> # A tibble: 6 x 14
#>   discarded_colum… discarded_rows qualifiers data  model variance variable
#>   <list>           <list>         <list>     <lis> <lis> <list>   <chr>   
#> 1 <tibble [32 × 0… <tibble [0 × … <tibble [… <tib… <prc… <tibble… C       
#> 2 <tibble [32 × 0… <tibble [0 × … <tibble [… <tib… <prc… <tibble… C/N     
#> 3 <tibble [32 × 0… <tibble [0 × … <tibble [… <tib… <prc… <tibble… Cu      
#> 4 <tibble [32 × 0… <tibble [0 × … <tibble [… <tib… <prc… <tibble… d13C    
#> 5 <tibble [32 × 0… <tibble [0 × … <tibble [… <tib… <prc… <tibble… d15N    
#> 6 <tibble [32 × 0… <tibble [0 × … <tibble [… <tib… <prc… <tibble… Ti      
#> # … with 7 more variables: PC1 <dbl>, PC2 <dbl>, PC3 <dbl>, PC4 <dbl>,
#> #   PC5 <dbl>, PC6 <dbl>, scores <list>

# scores, requalified
nested_pca %>% unnest(c(qualifiers, scores))

#> # A tibble: 32 x 15
#>    discarded_colum… discarded_rows depth zone  row_number data  model variance
#>    <list>           <list>         <dbl> <chr>      <int> <lis> <lis> <list>  
#>  1 <tibble [32 × 0… <tibble [0 × …  0.25 Zone…          1 <tib… <prc… <tibble…
#>  2 <tibble [32 × 0… <tibble [0 × …  0.75 Zone…          2 <tib… <prc… <tibble…
#>  3 <tibble [32 × 0… <tibble [0 × …  1.25 Zone…          3 <tib… <prc… <tibble…
#>  4 <tibble [32 × 0… <tibble [0 × …  1.75 Zone…          4 <tib… <prc… <tibble…
#>  5 <tibble [32 × 0… <tibble [0 × …  2.5  Zone…          5 <tib… <prc… <tibble…
#>  6 <tibble [32 × 0… <tibble [0 × …  3.5  Zone…          6 <tib… <prc… <tibble…
#>  7 <tibble [32 × 0… <tibble [0 × …  4.5  Zone…          7 <tib… <prc… <tibble…
#>  8 <tibble [32 × 0… <tibble [0 × …  5.5  Zone…          8 <tib… <prc… <tibble…
#>  9 <tibble [32 × 0… <tibble [0 × …  6.5  Zone…          9 <tib… <prc… <tibble…
#> 10 <tibble [32 × 0… <tibble [0 × …  7.5  Zone…         10 <tib… <prc… <tibble…
#> # … with 22 more rows, and 7 more variables: loadings <list>, PC1 <dbl>,
#> #   PC2 <dbl>, PC3 <dbl>, PC4 <dbl>, PC5 <dbl>, PC6 <dbl>