Alta Lake Lead-210 and Carbon-14 Ages

The alta_lake_210Pb_ages object contains raw depths and ages for Alta Lake (Whistler, British Columbia, Canada) core AL-GC2 (Dunnington et al. 2016; Dunnington 2015). For these values, ages were calculated from Lead-210 activities using the constant rate of supply (CRS) model (Appleby and Oldfield 1983). The alta_lake_14C_ages object contains one uncalibrated Carbon-14 measurement from the same core. The alta_lake_bacon_ages object contains the combined result of the Lead-210 and the Carbon-14 ages as modelled by the rbacon package (Blaauw and Christen 2011).

alta_lake_210Pb_ages

alta_lake_14C_ages

alta_lake_bacon_ages

alta_lake_geochem

Format

An object of class tbl_df (inherits from tbl, data.frame) with 9 rows and 5 columns.

An object of class tbl_df (inherits from tbl, data.frame) with 1 rows and 5 columns.

An object of class tbl_df (inherits from tbl, data.frame) with 58 rows and 5 columns.

An object of class tbl_df (inherits from tbl, data.frame) with 192 rows and 9 columns.

References

Appleby, P. G., and F. Oldfield. "The Assessment of 210Pb Data from Sites with Varying Sediment Accumulation Rates." Hydrobiologia 103, no. 1 (July 1, 1983): 29–35. doi:10.1007/BF00028424 .

Blaauw, Maarten, and J. Andrés Christen. "Flexible Paleoclimate Age-Depth Models Using an Autoregressive Gamma Process." Bayesian Analysis 6, no. 3 (September 2011): 457–74. doi:10.1214/ba/1339616472 .

Dunnington, Dewey W., Ian S. Spooner, Chris E. White, R. Jack Cornett, Dave Williamson, and Mike Nelson. "A Geochemical Perspective on the Impact of Development at Alta Lake, British Columbia, Canada." Journal of Paleolimnology 56, no. 4 (November 2016): 315–330. doi:10.1007/s10933-016-9919-x .

Dunnington, Dewey W. "A 500-Year Applied Paleolimnological Assessment of Environmental Change at Alta Lake, Whistler, British Columbia, Canada." M.Sc. Thesis, Acadia University, 2015. https://scholar.acadiau.ca/islandora/object/theses:411.

Examples

alta_lake_210Pb_ages
#> # A tibble: 9 × 5
#>   core                            depth_cm age_year_ad age_error_yr type        
#>   <chr>                              <dbl>       <dbl>        <dbl> <chr>       
#> 1 Dunnington et al. 2016 / AL-GC2      0         2015.         0    Sediment Le…
#> 2 Dunnington et al. 2016 / AL-GC2      1         2008          0.34 Sediment Le…
#> 3 Dunnington et al. 2016 / AL-GC2      1.5       2003.         0.56 Sediment Le…
#> 4 Dunnington et al. 2016 / AL-GC2      2         1998.         0.86 Sediment Le…
#> 5 Dunnington et al. 2016 / AL-GC2      3         1982.         2.25 Sediment Le…
#> 6 Dunnington et al. 2016 / AL-GC2      4         1966.         4.73 Sediment Le…
#> 7 Dunnington et al. 2016 / AL-GC2      5         1947.         8.21 Sediment Le…
#> 8 Dunnington et al. 2016 / AL-GC2      6         1922.        27.0  Sediment Le…
#> 9 Dunnington et al. 2016 / AL-GC2      7         1896         57.2  Sediment Le…
alta_lake_14C_ages
#> # A tibble: 1 × 5
#>   core                            depth_cm age_14C age_error_14C type           
#>   <chr>                              <dbl>   <dbl>         <dbl> <chr>          
#> 1 Dunnington et al. 2015 / AL-GC2     29.5     340            30 Carbon-14 age …
alta_lake_bacon_ages
#> # A tibble: 58 × 5
#>    depth_cm age_min_year_BP age_max_year_BP age_median_year_BP age_weighted_me…¹
#>  *    <dbl>           <dbl>           <dbl>              <dbl>             <dbl>
#>  1      1            -59.1           -57.1              -58.1             -58.1 
#>  2      1.5          -54.3           -52.4              -53.3             -53.3 
#>  3      2            -50.5           -46.7              -48.6             -48.6 
#>  4      2.5          -43.9           -37.1              -40.5             -40.5 
#>  5      3            -39.3           -25.8              -32.4             -32.4 
#>  6      3.5          -30.9           -17.8              -24.5             -24.4 
#>  7      4            -26.8            -4.42             -16.7             -16.3 
#>  8      4.5          -18.0             3.28              -7.79             -7.61
#>  9      5            -13.6            17.8                0.46              1.01
#> 10      5.5           -5.91           29.0                9.8              10.4 
#> # … with 48 more rows, and abbreviated variable name ¹​age_weighted_mean_year_BP