Halifax lakes water chemistry and top/bottom diatom counts

A subset of well-labeled, clean diatom count data for 44 Halifax-area (Nova Scotia) lakes, an analysis of which has been published by Ginn et al. (2015).

halifax_lakes_plottable

Format

An object of class tbl_df (inherits from tbl, data.frame) with 114 rows and 5 columns.

Source

Neotoma paleoecology database (https://neotomadb.org)

References

Ginn, Brian K., Thiyake Rajaratnam, Brian F. Cumming, and John P. Smol. "Establishing Realistic Management Objectives for Urban Lakes Using Paleolimnological Techniques: An Example from Halifax Region (Nova Scotia, Canada)." Lake and Reservoir Management 31, no. 2 (April 3, 2015): 92-108. doi: 10.1080/10402381.2015.1013648 .

Examples

halifax_lakes_plottable
#> # A tibble: 114 × 5
#>    location      sample_type taxon                             count rel_abund
#>    <chr>         <chr>       <fct>                             <dbl>     <dbl>
#>  1 Anderson Lake bottom      Aulacoseira distans                26       4.65 
#>  2 Anderson Lake bottom      Eunotia exigua                     13.5     2.42 
#>  3 Anderson Lake bottom      Fragilariforma exigua               6.5     1.16 
#>  4 Anderson Lake bottom      Tabellaria fenestrata              23       4.11 
#>  5 Anderson Lake bottom      Tabellaria flocculosa (strain IV)  21       3.76 
#>  6 Anderson Lake bottom      Other                             469      83.9  
#>  7 Anderson Lake top         Aulacoseira distans                 8.5     1.87 
#>  8 Anderson Lake top         Eunotia exigua                      0       0    
#>  9 Anderson Lake top         Fragilariforma exigua               1.5     0.330
#> 10 Anderson Lake top         Tabellaria fenestrata              26       5.71 
#> # … with 104 more rows