Keji lakes core diatom counts

A subset of well-labeled, clean diatom count data for 3 Keji-area (Nova Scotia) lakes, which form part of the analysis in Ginn et al. (2007). These data are in the form of a mudata object, from which location information, parameter information, and data can be extracted using tbl_locations, tbl_params, and tbl_data (respectively). Also, a subset of these data amenable to plotting have been made available.

keji_lakes

keji_lakes_plottable

Format

An object of class mudata (inherits from list) of length 5.

An object of class tbl_df (inherits from tbl, data.frame) with 202 rows and 5 columns.

Source

Neotoma paleoecology database (http://neotomadb.org/)

References

Ginn, Brian K., Brian F. Cumming, and John P. Smol. "Long-Term Lake Acidification Trends in High- and Low-Sulphate Deposition Regions from Nova Scotia, Canada." Hydrobiologia 586, no. 1 (July 1, 2007): 261-75. https://doi.org/10.1007/s10750-007-0644-3.

Examples

library(mudata2)
data(keji_lakes)

keji_lakes %>%
  tbl_data()

#> # A tibble: 6,363 x 9
#>    dataset location param  depth value neotoma_sample_… unit_name
#>    <chr>   <chr>    <chr>  <dbl> <dbl>            <int> <chr>    
#>  1 neotom… Peskawa… Acti…  0.125     1           166528 Peskawa …
#>  2 neotom… Peskawa… Acti…  0.375     0           166529 Peskawa …
#>  3 neotom… Peskawa… Acti…  2.38      2           166530 Peskawa …
#>  4 neotom… Peskawa… Acti…  3.82      1           166531 Peskawa …
#>  5 neotom… Peskawa… Acti…  5.12      0           166532 Peskawa …
#>  6 neotom… Peskawa… Acti…  6.38      1           166533 Peskawa …
#>  7 neotom… Peskawa… Acti…  7.38      1           166534 Peskawa …
#>  8 neotom… Peskawa… Acti…  8.38      0           166535 Peskawa …
#>  9 neotom… Peskawa… Acti…  9.38      2           166536 Peskawa …
#> 10 neotom… Peskawa… Acti… 10.1       0           166537 Peskawa …
#> # … with 6,353 more rows, and 2 more variables: neotoma_dataset_id <dbl>,
#> #   variable_units <chr>