Long Lake Carbon-14 Ages

This object contains several uncalibrated Carbon-14 measurements from Long Lake (Nova Scotia-New Brunswick Border Region, Canada) core LL-PC2 (Dunnington et al. 2017; White 2012). The long_lake_bacon_ages object contains the result of the Carbon-14 ages as modelled by the rbacon package (Blaauw and Christen 2011).

long_lake_14C_ages

long_lake_bacon_ages

long_lake_plottable

Format

An object of class tbl_df (inherits from tbl, data.frame) with 5 rows and 5 columns.

An object of class tbl_df (inherits from tbl, data.frame) with 86 rows and 5 columns.

An object of class tbl_df (inherits from tbl, data.frame) with 51 rows and 10 columns.

References

Blaauw, Maarten, and J. Andrés Christen. "Flexible Paleoclimate Age-Depth Models Using an Autoregressive Gamma Process." Bayesian Analysis 6, no. 3 (September 2011): 457–74. doi: 10.1214/ba/1339616472 .

Dunnington, Dewey W., Hilary White, Ian S. Spooner, Mark L. Mallory, Chris White, Nelson J. O’Driscoll, and Nic R. McLellan. "A Paleolimnological Archive of Metal Sequestration and Release in the Cumberland Basin Marshes, Atlantic Canada." FACETS 2, no. 1 (May 23, 2017): 440–60. doi: 10.1139/facets-2017-0004 .

White, Hilary E. "Paleolimnological Records of Post-Glacial Lake and Wetland Evolution from the Isthmus of Chignecto Region, Eastern Canada." M.Sc. Thesis, Acadia University, 2012. https://scholar.acadiau.ca/islandora/object/theses:247.

Examples

long_lake_14C_ages

#> # A tibble: 5 x 5
#>   core                     depth_cm age_14C age_error_14C type                  
#>   <chr>                       <dbl>   <dbl>         <dbl> <chr>                 
#> 1 Dunnington et al. 2017 …      9       623            34 Carbon-14 age / twig …
#> 2 Dunnington et al. 2017 …     37.3     575            18 Carbon-14 age / plant 
#> 3 Dunnington et al. 2017 …     46      4396            55 Carbon-14 age / wood …
#> 4 Dunnington et al. 2017 …     70.5    8482            28 Carbon-14 age / wood  
#> 5 Dunnington et al. 2017 …     94      9148            49 Carbon-14 age / wood …

long_lake_bacon_ages

#> # A tibble: 86 x 5
#>    depth   min   max median wmean
#>    <dbl> <dbl> <dbl>  <dbl> <dbl>
#>  1     9  536.  667    598.  603.
#>  2    10  572.  942.   695.  714.
#>  3    11  583. 1263.   775.  822.
#>  4    12  637. 1378.   893.  928.
#>  5    13  669. 1580.   996. 1034.
#>  6    14  738. 1703.  1112. 1144.
#>  7    15  787. 1922   1218. 1254.
#>  8    16  870. 2031.  1331. 1365.
#>  9    17  925. 2218.  1443. 1478.
#> 10    18 1014. 2329.  1556. 1589.
#> # … with 76 more rows

long_lake_plottable

#> # A tibble: 51 x 10
#>    dataset     location param depth value     sd units zone       n n_detect
#>    <chr>       <chr>    <chr> <dbl> <dbl>  <dbl> <chr> <chr>  <int>    <int>
#>  1 long_lake17 LL PC2   C/N       1  15.4 NA     NA    Unit 3     1        1
#>  2 long_lake17 LL PC2   C/N       6  15.8  0.290 NA    Unit 3     2        2
#>  3 long_lake17 LL PC2   C/N      11  24.0 NA     NA    Unit 2     1        1
#>  4 long_lake17 LL PC2   C/N      16  21.4 NA     NA    Unit 2     1        1
#>  5 long_lake17 LL PC2   C/N      21  23.2 NA     NA    Unit 2     1        1
#>  6 long_lake17 LL PC2   C/N      26  25.4 NA     NA    Unit 2     1        1
#>  7 long_lake17 LL PC2   C/N      43  27.2 NA     NA    Unit 2     1        1
#>  8 long_lake17 LL PC2   C/N      48  24.3 NA     NA    Unit 2     1        1
#>  9 long_lake17 LL PC2   C/N      53  30.2 NA     NA    Unit 2     1        1
#> 10 long_lake17 LL PC2   C/N      63  24.4 NA     NA    Unit 2     1        1
#> # … with 41 more rows